Implements a function re-assigning voxels with a given label to their closest regions [BBPP42-495]
* The function which re-assigns voxels is deterministic.
* This function is sensitive to the order with which re-assignable voxels are visited.
* The utils function that picks a non-void closest voxel with different label has a time complexity
  in O(R * V * log N) where R is the number of voxels to re-assign, V is the total number
  of voxels and N is the largest dimension of the 3D image.
* In practical applications, we have: R << V and the bound is closer to O(R^2 * log N)

As suggested by the reviewer, this implementation could be done with a
a scipy KD-tree (https://en.wikipedia.org/wiki/K-d_tree). But this need
to be tested as it may require much more memory depending on the input (res. of 10mu --> +5GB).
Still, a KD-tree is likely to be more efficient and would be insensitive to the ordering
of vertices. To be implemented as an alernative algorithm in a subsequent commit.

Change-Id: Ie7f98d8c64ae98cea5c804c2768e4feda0ecfc27