Uses scipy's nearest-neighbor algorithm as an atlernative for assigning voxels to their closest regions [BBPP-495]
* If the 'algorithm' parameter equals to 'nearest-neighbor'(default), assign_to_closest_region()
will use scipy's nearest-neighbor algorithm (relies in turn on cKDTree).
* This method has the advantage to be standard and not to depend on any visit order of the
voxels to be interpolated.
* If 'algorithm' is set to 'competitive-nearest-neighbor', then neighboring regions compete to have the largest
share of voxels during re-assignment. Although the process is still deterministic,
the order with which re-assignable voxels are visited does change the result.
This method has the advantage to keep the memory requirement low and the search for
nearest neighbors is done in average in very small neighborhoods.
* If the user-specified algorithm is none of the two above values, the function
raises an AtlasAnalysisError.
In this commit, the AtlasError exception has been rename as AtlasErrorAnalysis
so as to be reused in the context of curation and possibly other contexts.
Change-Id: Iddc39ae362fb0dc7c7eaa0511505e20b8e65f954