Implements a reporting function for voxel counts [BBPP42-565]
* A new module named reporting.py has been created.
* The main report function writes a report under the form of a JSON file.
(This function sits in app/reporting.py).
A report instance can also be loaded back from a JSON file.
* The main report includes the voxel count of each region (or voxel label) of the input nrrd file.
The input nrrd file is assumed to hold a 3D image, e.g., an atlas annotation file.
* If requested (optional), the report indicates whether the 3D image, or its leaf regions,
are connected or not. The report contains an histogram of the connected component counts
of the whole 3D image and a similar  histogram for each leaf region.
The bins of the histogram are of the form [10**i, 10**(i + 1)) with i <= 7. The first bin is
[0, 10). To each bin correponds the number of connected components whose size (number of voxels)
lies within the bounds of the bin.
* If requested (optional), the report indicates if there are cavities in the whole image and
its regions as well. The voxel counts of cavities are appended in the 'cavities'
section of the report.
* In addition, the report comprises the number of regions and the list of region ids.
It contains also the image dimensions, its offset and the voxel dimensions.
* Internally, the report is an instance of VoxelCountReport, a class with methods such
as from_voxel_data(), to_dict(), from_dict() and load(). This class is built up from
smaller report classes such as ConnectivityReport and CavityReport with similar
functionalities.

Change-Id: Icca85c80a2042522c3718635642c68d8cd735e7f