Adds a curation function to create a uniform black margin around a 3D image [BBPP42-462]
* This feature comes in handy when generating a boundary mesh from 3D volumetric file
if this file has been cropped to its smallest enclosing box. Indeed, such a mesh may
have holes because some voxels of the volume boundary are also in the boundary of
the image.

* The function adds the same number of black voxels in each dimension.
Hence each dimension is increased by 2 * margin where margin is the user input
parameter.

* The VoxelData offset is adjusted accordingly and the output object is written
to a file bearing the name specified by the user.

Change-Id: Ide71ca9089b7013436d42b4f4a48bf7e58193f27