CLI for the display of flatmap images and flatmap volume histograms [NETSEM-972]
These changes comprise

* The command line 'atlas-analysis visualization show-flatmap-image', a command to show the flatmap image
or a downscaled version of it.

* The command line 'atlas-analysis visualization show-flatmap-histogram', a command to show the flatmap fiber
volumes or a downscaled version of it. The fiber over a pixel
is the set of voxels which are mapped to this pixel.
The fiber volume is the number of these voxels.
The displayed histogram is an elevation grid (3D) lying
above the flatmap image in the (x,y)-plane.

* Downscaling is specified by providing a target resolution, i.e.,
a target width to set the x-dimension of the flatmap image. This is necessary
when flatmaps are large (>35'000'000 pixels). The aspect ratio is preserved,
or close to be preserved.
The numpy-based scikit-image module is used under the hood.

* The visualization relies on plotly.graph_objects. So, the rendering is done
in the browser where manipulators are available (zoom in and out, rotations).
The images can be saved as png files from the browser.

Change-Id: I8d7e6f1fc344d98944b1ea6ad782dcdef886782f