Command line interface to crop nrrd files with the smallest AABB [BBPP82-310]
* Addition of the command `atlas-analysis curation NRRD -o OUTPUTPATH -m MARGIN`
This command crops the volume enclosed by the file NRRD with the smallest
Axis Aligned Bounding Box (AABB) containing it and saves it to OUTPUTPATH.

* Optionally, some margin can be added around the cropped volume.
This can be useful to workaround the visualization artefact of itk-snap
or to prevent Ultraliser from crashing.

* The NRRD header offset is modified accordingly.

* The implementation is not new but relies now essentially on the tools of
voxcell.math_utils.

This feature allows to handle much smaller nrrd files than
than whole brain atlases.

Change-Id: I61763f5840a6cdb9698ee4725cd4817f0bd4111d