A function to interpolate N-D vector fields.
This change implements a function that interpolates in place
a vector field represented as a numpy numeric array.

This is a simple wrapper around scipy interpolators.

It provides a natural interface to interpolate:

* direction vector fields,
* orientation fields,
* distance fields,

where values are unknown (if computations failed) or
known to be inaccurate.
These vector fields arise in the context of the Atlas Building Pipeline.

Change-Id: I53d7974f750b777c3451a8630fa479fb35f6266c