[FIX] Fix the full resolution flatmap image and improve a warning message
In order to be displayed properly, the flatmap image (boolean array)
must be cast to the type int.
This cast operation was missing when the full flatmap image resolution
is requested.

In addition, we modified the response of the script when a user requests
accidentally an upscale of the flatmap image. In this case, a warning is
issued and the full resolution is used.

The warning describes with more details why the specified resolution parameter
is problematic.

Change-Id: Ie8c2dbb83ed8f7bd28be6490f0e381c6e40e407e