atlas-analysis planes, coordinates, reporting: minor enhancements [NSETM-1139][FMHIPP-50]
This change implements three enhancements:

* [FMIHPP-50] atlas-analysis coordinates: the user is now allowed to specify the
identifiers of the output nrrd file
* [FMIHPP-50] atlas-analysis reporting voxel-count-report: in addition to the list
of all region identifiers, the report
contains the mapping to region acronyms and names, provided a
hierarchy json file is given as input (optional) [FMHIPP-50]
* [NSETM-1139] atlas-analysis planes: defines an alternative output format for the
planes, that is the equation format [x, y, z, A, B, C, D] where
(x, y, z) is the intersection point of the centerline with the plane
and A * X + B * Y + C * Z = D is an equation defining the plane.
This change is backward compatible and the `atlas-analysis planes draw`
CLI is able to read any of the two formats (quaternion or equation).

The first enhancement is a follow-up on the work on the segmentation of
Mouse CA1 Hippocampus.

The second enhencement stems out of a discussion with Davide Bella in
the same context.

The last enhancement stems out of a discussion with Alexis Arnaudon in the context
of [NSETM-1139]. Both output plane formats have been documented in the CLI help message.
Some utils functions located originally in atlas_building_tools/planes.py have been moved to
atlas_building_tools/planes/maths.py.

Note that atlas-analysis uses only one format internally, the quaternionic format.
The equation format is only exposed as an output format.

Some utils functions originally located in atlas_building_tools/maths.py have been moved to
atlas_building_tools/planes/maths.py because they depend on the choice of ZVECTOR as reference
vector to build quaternions, an 'internal' process only defined in planes.py. (placement-algorithm
uses for instance YVECTOR instead.)

Note: Alexis Arnaudon and Thomas Delemontex rightly observed that adding a new format
adds confusion. The code will be refactored in a forthcoming commit by means of
a single Plane class providing all the necessary accessors.

Change-Id: I0e0d4c3546f411e292f8199c32eb77b21b753aea