Refactor the plane module using a Plane class [NSETM-1139]
This change originates from the comments of Alexis Arnaudon and
Thomas Delemontex made during the code review
https://bbpcode.epfl.ch/code/#/c/49738/.

The change implements a Plane class whose underlying abstract data
is an oriented pointed plane of the Euclidean 3D space, i.e.,
an affine hyperplane defined through a distinguished 3D point
or 'anchor' point and a distinguished unit normal vector (distinguished
among the two possiblities).

The anchor point comes in handy when rendering the plane in 3D.

The Plane constructor takes as arguments the anchor point coordinates
[x, y, z] and a unit vector [A, B, C] orthogonal to the plane:
the plane is the set of points [X, Y, Z] satisfying
A * X + B * Y + C * Z = D and where D = A * x + B * y + C * z where
[x, y, z] is the anchor point.

The Plane class offers a method which returns a direct orthonomal
basis of the plane (direct with respect to the normal vector [A, B, C]).
This is useful to draw the plane and it should be useful to
draw on the plane too (Alexis Arnaudon use case).

Change-Id: I4da633593d5137b21a0bf3db76deb2a5f530938d