A function returning an orthonormal basis of a plane with some optimal alignment property [NSETM-1170]
This change implements
* the method get_best_alignment_basis of the planes.maths.Plane class.
* the method create_orthogonal_planes in the planes.maths module

An orthonormal basis of a plane with some best alignment property
-----------------------------------------------------------------
Given a target non-zero 3D vector d, interpreted as a direction, the function returns
a direct orthonormal basis of the plane whose first vector follows the direction d
as closely as possible, i.e., it maximizes the scalar product v dot d  among all
basis vectors v of the plane.

If the target direction is the direction of the unit normal vector of the given plane,
the function returns get_basis(). Note that in this case any basis is optimal.

A function returning a sequence of planes orthogonal to a 3D line segment
-------------------------------------------------------------------------
Given a 3D line segment [start, end] and a list of N float numbers called `positions`,
the function planes.maths.create_orthogonal_planes returns N Plane objects
whose normal unit vector is (end - start) / || end - start || and whose anchor points
are distributed along the segment [start, end] according the specified numbers.
The input number list is transformed by an affine map to fit in [0.0, 1.0].
For instance, if [1.0, 2.0, 4.0] is the input number list for anchor positions,
the generated anchors are start, start + (end - start) / 3.0 and end.

In addition,

* the functions planes._create_planes and planes._create_centerline have been
made public, as requested by Alexis Arnaudon.

* the output format called 'equation' format has been removed and is now superseded by
the 'point_normal' format, the natural format to describe an affine and oriented pointed
plane in the 3D Euclidean space: a 6D float vector [x, y, z, A, B, C] where [x, y, z]
are the coordinates of a favoured point of the plan and [A, B, C] is a unit normal
vector of the plane.

Change-Id: Icd3867869b3aacef900fef220d6fa7d373792eab