Implements a consistency check for VoxelData metadata [NSETM-1035]
The added function assert_meta_properties checks that a
list of VoxelData objects share the same
* VoxelData.shape
* VoxelData.voxel_dimensions
* VoxelData.offset

and raises otherwise.

This allows for instance to check that VoxelData objects
corresponding to annotations and direction vectors have the
same underlying brain region (3D volume with fixed shape)
and the same coordinate system (voxel dimensions and offset).

This is written in the context of NSETM-1035 because the change
https://bbpcode.epfl.ch/code/#/c/50053/ in nse/atlas-building-tools
will use this function.

Change-Id: I0e9353920860922fea319c53e7b39d13fba93248